changefeedccl: Move topic related code to topic.
Pure code move refactoring to move topic related code
to topic file.

Release Notes: None
